Various Events Page changes
-date & time: a clock icon should be in front of the time, just like the
calendar in front of the date. You can make the font of the date the same
as "women", "men" and the region instead of just bold.

-location: the region is not showing any more, probably due to the changed
you made with the attribute?

-Participants: a subtle divider between the two sexes would be nice. Just
like the dividers used in the logged in menu.

-mobile: make sure the participants are next to each other (like desktop
version) instead of below each other.will take less space.

-location: same remark as location page, different mobile and desktop
version. You can replace the more details button with just a bold text
like we have on the location page right below bottom, takes less space and
is cleaner.

-bookingsform: Ive noticed the form area is bigger then the details or
left boxes area. Is this on purpose or necessary for all the information
we need to collect? Thinking out loud.

-bookingsform: no need to show the ticket spaces. we use the
green/orange/red phrases in the participants box. They don't need to know
the exact spaces. Will make up room, we need the space.

-similar events like this: I would like to have it the same format/design
as on the location page (box)

-similar events like this: more details buttons just like location, bigger

-similar events like this mobile: same remarks as location.

-event description and image mobile: completely aligned to the left side,
should be more to the right llike all the other content (see attachment)