Events Page date and time changes
-date & time: You can make the font of the date the same as "women", "men"
and the region. This way all the blocks have 1 sort of highlight. it's
even a better idea to remove the clock (my mistake, sorry), this way they
all look the same. Like this:

[calendar icon] 7th of august, 2014 (font like "women")
from 7:00 pm until 10:00 pm [info icon with hoover info box that mentions
"we advise you to come 30 min earlier"]