Events Lists tweaks
-in tables we can just use "07/10/2014", "7:00pm - 10:00pm" and both of the
icons to save space.

-Mobile version events table: Well done to remove the thumbnail in the
events tables in the mobile version, looks a lot cleaner! I would also
remove the "more details" button/column to make more space. This way we
can make (more or less) sure that the date, hour and ages are on one line.

-events tables: a little bit more distance between the event title and the
location below it so it's clear both are not together. Maybe also a
slightly different grey for the region to make it stand out a bit more.